aeval/registry/requirements.txt
fastapi>=0.115
uvicorn>=0.34
pyyaml>=6.0
pydantic>=2.0
httpx>=0.27
